+++ cli/pcluster/cfnconfig.py
@@ -360,12 +360,6 @@ class ParallelClusterConfig(object):
                     "VPC section [%s] used in [%s] section is not defined" % (vpc_section, self.__cluster_section)
                 )
 
-        # Check that cidr and public ips are not both set
-        cidr_value = self.__config.get(vpc_section, "compute_subnet_cidr", fallback=None)
-        public_ips = self.__config.getboolean(vpc_section, "use_public_ips", fallback=True)
-        if self.__sanity_check:
-            ResourceValidator.validate_vpc_coherence(cidr_value, public_ips)
-
     def __check_account_capacity(self):
         """Try to launch the requested number of instances to verify Account limits."""
         if self.parameters.get("Scheduler") == "awsbatch" or self.parameters.get("ClusterType", "ondemand") == "spot":