+++ cli/pcluster/config_sanity.py
@@ -47,17 +47,6 @@ class ResourceValidator(object):
         return "aws"
 
     @staticmethod
-    def validate_vpc_coherence(cidr_value, public_ip):
-        """
-        Check that cidr_value and public_ip parameters are not conflicting.
-
-        :param cidr_value: the value of compute_subnet_cidr set by the user (default should be None)
-        :param public_ip: the value of use_public_ips set by the user (default should be True)
-        """
-        if cidr_value and public_ip is False:
-            ResourceValidator.__fail("VPC COHERENCE", "compute_subnet_cidr needs use_public_ips to be true")
-
-    @staticmethod
     def __check_sg_rules_for_port(rule, port_to_check):
         """
         Verify if the security group rule accepts connections on the given port.
@@ -355,7 +344,8 @@ class ResourceValidator(object):
                     ),
                     (
                         ["cloudformation:DescribeStacks"],
-                        "arn:%s:cloudformation:%s:%s:stack/parallelcluster-*" % (partition, self.region, account_id),
+                        ["cloudformation:DescribeStackResource"],
+                        "arn:%s:cloudformation:%s:%s:stack/parallelcluster-*/*" % (partition, self.region, account_id),
                     ),
                     (["s3:GetObject"], "arn:%s:s3:::%s-aws-parallelcluster/*" % (partition, self.region)),
                     (["sqs:ListQueues"], "*"),
@@ -438,7 +428,7 @@ class ResourceValidator(object):
                 self.__fail(resource_type, e.response.get("Error").get("Message"))
         # EC2 Placement Group
         elif resource_type == "EC2PlacementGroup":
-            if resource_value == "DYNAMIC":
+            if resource_value == "DYNAMIC" or resource_value == "NONE":
                 pass
             else:
                 try: