+++ tests/integration-tests/tests/common/schedulers_common.py
@@ -206,11 +206,17 @@ class SlurmCommands(SchedulerCommands):
     def __init__(self, remote_command_executor):
         super().__init__(remote_command_executor)
 
-    @retry(retry_on_result=lambda result: result == "Unknown", wait_fixed=seconds(7), stop_max_delay=minutes(5))
+    @retry(
+        retry_on_result=lambda result: "JobState" not in result
+        or any(value in result for value in ["EndTime=Unknown", "JobState=RUNNING", "JobState=COMPLETING"]),
+        wait_fixed=seconds(7),
+        stop_max_delay=minutes(5),
+    )
     def wait_job_completed(self, job_id):  # noqa: D102
-        result = self._remote_command_executor.run_remote_command("scontrol show jobs -o {0}".format(job_id))
-        match = re.search(r"EndTime=(.+?) ", result.stdout)
-        return match.group(1)
+        result = self._remote_command_executor.run_remote_command(
+            "scontrol show jobs -o {0}".format(job_id), raise_on_error=False
+        )
+        return result.stdout
 
     def get_job_exit_status(self, job_id):  # noqa: D102
         result = self._remote_command_executor.run_remote_command("scontrol show jobs -o {0}".format(job_id))
@@ -242,7 +248,7 @@ class SlurmCommands(SchedulerCommands):
         if slots:
             submission_command += " -n {0}".format(slots)
         if nodes > 1:
-            submission_command += " -N {0}".format(slots)
+            submission_command += " -N {0}".format(nodes)
         submission_command += " {1}".format(nodes, script_name)
         return self._remote_command_executor.run_remote_command(submission_command, additional_files=additional_files)
 
@@ -268,23 +274,39 @@ class TorqueCommands(SchedulerCommands):
     def __init__(self, remote_command_executor):
         super().__init__(remote_command_executor)
 
+    @retry(
+        retry_on_result=lambda result: "job_state = C" not in result, wait_fixed=seconds(7), stop_max_delay=minutes(5)
+    )
     def wait_job_completed(self, job_id):  # noqa: D102
-        raise NotImplementedError
+        result = self._remote_command_executor.run_remote_command("qstat -f {0}".format(job_id))
+        return result.stdout
 
     def get_job_exit_status(self, job_id):  # noqa: D102
-        raise NotImplementedError
+        result = self._remote_command_executor.run_remote_command("qstat -f {0}".format(job_id))
+        match = re.search(r"exit_status = (\d+)", result.stdout)
+        return match.group(1)
 
     def assert_job_submitted(self, qsub_output):  # noqa: D102
-        raise NotImplementedError
+        __tracebackhide__ = True
+        # qsub_output is the id of the job in case of successful submissions
+        id = qsub_output
+        # check that the job exists
+        self._remote_command_executor.run_remote_command("qstat -f {0}".format(id))
+        return id
 
-    def submit_command(self, command):  # noqa: D102
-        raise NotImplementedError
+    def submit_command(self, command, nodes=1, slots=None):  # noqa: D102
+        flags = "-l nodes={0}:ppn={1}".format(nodes or 1, slots or 1)
+        return self._remote_command_executor.run_remote_command(
+            "echo '{0}' | qsub {1}".format(command, flags), raise_on_error=False
+        )
 
     def submit_script(self, script, nodes=1):  # noqa: D102
         raise NotImplementedError
 
     def assert_job_succeeded(self, job_id, children_number=0):  # noqa: D102
-        raise NotImplementedError
+        __tracebackhide__ = True
+        status = self.get_job_exit_status(job_id)
+        assert_that(status).is_equal_to("0")
 
     def compute_nodes_count(self):  # noqa: D102
         result = self._remote_command_executor.run_remote_command(